requirements.txt
Django>=5.0,<6.0
gunicorn>=21.2.0
whitenoise>=6.5.0
psycopg2-binary>=2.9.9
django-unfold>=0.40.0
openai>=1.0.0
stripe>=7.0.0
Pillow>=10.0.0
packaging>=24.0
